Remove duplicate validation functions from conditions.go
