+++ utils/conditions/rollouts_test.go
@@ -1,15 +1,12 @@
 package conditions
 
 import (
-	"fmt"
-	"math"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/kubernetes/pkg/controller"
 	"k8s.io/utils/pointer"
 
@@ -225,231 +222,6 @@ func TestRemoveCondition(t *testing.T) {
 			assert.Equal(t, test.status, test.expectedStatus)
 		})
 	}
-}
-
-func TestVerifyRolloutSpecBlueGreen(t *testing.T) {
-	validRollout := &v1alpha1.Rollout{
-		Spec: v1alpha1.RolloutSpec{
-			Selector: &metav1.LabelSelector{
-				MatchLabels: map[string]string{"key": "value"},
-			},
-			Strategy: v1alpha1.RolloutStrategy{
-				BlueGreen: &v1alpha1.BlueGreenStrategy{
-					PreviewService: "preview",
-					ActiveService:  "active",
-				},
-			},
-		},
-	}
-	assert.Nil(t, VerifyRolloutSpec(validRollout, nil))
-
-	sameSvcs := validRollout.DeepCopy()
-	sameSvcs.Spec.Strategy.BlueGreen.ActiveService = "preview"
-	sameSvcsCond := VerifyRolloutSpec(sameSvcs, nil)
-	assert.NotNil(t, sameSvcsCond)
-	assert.Equal(t, DuplicatedServicesMessage, sameSvcsCond.Message)
-	assert.Equal(t, InvalidSpecReason, sameSvcsCond.Reason)
-
-	scaleLimitLargerThanRevision := validRollout.DeepCopy()
-	scaleLimitLargerThanRevision.Spec.Strategy.BlueGreen.ScaleDownDelayRevisionLimit = pointer.Int32Ptr(100)
-	scaleLimitLargerThanRevisionCond := VerifyRolloutSpec(scaleLimitLargerThanRevision, nil)
-	assert.NotNil(t, scaleLimitLargerThanRevisionCond)
-	assert.Equal(t, ScaleDownLimitLargerThanRevisionLimit, scaleLimitLargerThanRevisionCond.Message)
-	assert.Equal(t, InvalidSpecReason, sameSvcsCond.Reason)
-}
-
-func TestVerifyRolloutSpecBaseCases(t *testing.T) {
-	ro := &v1alpha1.Rollout{
-		Spec: v1alpha1.RolloutSpec{
-			Selector: &metav1.LabelSelector{
-				MatchLabels: map[string]string{"key": "value"},
-			},
-			Strategy: v1alpha1.RolloutStrategy{
-				Canary: &v1alpha1.CanaryStrategy{},
-				BlueGreen: &v1alpha1.BlueGreenStrategy{
-					ActiveService: "active",
-				},
-			},
-		},
-	}
-	cond := VerifyRolloutSpec(ro, nil)
-	assert.Equal(t, v1alpha1.InvalidSpec, cond.Type)
-	assert.Equal(t, InvalidSpecReason, cond.Reason)
-	assert.Equal(t, InvalidStrategyMessage, cond.Message)
-
-	validRollout := ro.DeepCopy()
-	validRollout.Spec.Strategy.Canary = nil
-	validRolloutCond := VerifyRolloutSpec(validRollout, nil)
-	assert.Nil(t, validRolloutCond)
-
-	selectorEverything := validRollout.DeepCopy()
-	selectorEverything.Spec.Selector = &metav1.LabelSelector{}
-	selectorEverythingConf := VerifyRolloutSpec(selectorEverything, nil)
-	assert.NotNil(t, selectorEverythingConf)
-	assert.Equal(t, RolloutSelectAllMessage, selectorEverythingConf.Message)
-	assert.Equal(t, InvalidSpecReason, selectorEverythingConf.Reason)
-
-	noSelector := validRollout.DeepCopy()
-	noSelector.Spec.Selector = nil
-	noSelectorCond := VerifyRolloutSpec(noSelector, nil)
-	assert.NotNil(t, noSelectorCond)
-	assert.Equal(t, fmt.Sprintf(MissingFieldMessage, ".Spec.Selector"), noSelectorCond.Message)
-	assert.Equal(t, InvalidSpecReason, noSelectorCond.Reason)
-
-	minReadyLongerThanProgessDeadline := validRollout.DeepCopy()
-	minReadyLongerThanProgessDeadline.Spec.MinReadySeconds = 1000
-	minReadyLongerThanProgessDeadlineCond := VerifyRolloutSpec(minReadyLongerThanProgessDeadline, nil)
-	assert.NotNil(t, minReadyLongerThanProgessDeadlineCond)
-	assert.Equal(t, InvalidSpecReason, minReadyLongerThanProgessDeadlineCond.Reason)
-	assert.Equal(t, RolloutMinReadyLongerThanDeadlineMessage, minReadyLongerThanProgessDeadlineCond.Message)
-}
-
-func TestVerifyRolloutSpecCanary(t *testing.T) {
-	zero := intstr.FromInt(0)
-	tests := []struct {
-		name           string
-		maxUnavailable *intstr.IntOrString
-		maxSurge       *intstr.IntOrString
-		steps          []v1alpha1.CanaryStep
-
-		notValid bool
-		reason   string
-		message  string
-	}{
-		{
-			name:           "Max Surge and Max Unavailable set to zero",
-			maxUnavailable: &zero,
-			maxSurge:       &zero,
-
-			notValid: true,
-			reason:   InvalidSpecReason,
-			message:  InvalidMaxSurgeMaxUnavailable,
-		},
-		{
-			name: "setWeight and pause both set",
-			steps: []v1alpha1.CanaryStep{{
-				Pause:     &v1alpha1.RolloutPause{},
-				SetWeight: pointer.Int32Ptr(10),
-			}},
-
-			notValid: true,
-			reason:   InvalidSpecReason,
-			message:  InvalidStepMessage,
-		},
-		{
-			name:  "experiment, setWeight, and pause are not set",
-			steps: []v1alpha1.CanaryStep{{}},
-
-			notValid: true,
-			reason:   InvalidSpecReason,
-			message:  InvalidStepMessage,
-		},
-		{
-			name: "setWeight over 0",
-			steps: []v1alpha1.CanaryStep{{
-				SetWeight: pointer.Int32Ptr(-1),
-			}},
-
-			notValid: true,
-			reason:   InvalidSpecReason,
-			message:  InvalidSetWeightMessage,
-		},
-		{
-			name: "setWeight less than 100",
-			steps: []v1alpha1.CanaryStep{{
-				SetWeight: pointer.Int32Ptr(110),
-			}},
-
-			notValid: true,
-			reason:   InvalidSpecReason,
-			message:  InvalidSetWeightMessage,
-		},
-		{
-			name: "Pause duration is not less than 0",
-			steps: []v1alpha1.CanaryStep{{
-				Pause: &v1alpha1.RolloutPause{
-					Duration: v1alpha1.DurationFromInt(-1),
-				},
-			}},
-
-			notValid: true,
-			reason:   InvalidSpecReason,
-			message:  InvalidDurationMessage,
-		},
-		{
-			name: "Pause duration invalid unit",
-			steps: []v1alpha1.CanaryStep{{
-				Pause: &v1alpha1.RolloutPause{
-					Duration: v1alpha1.DurationFromString("10z"),
-				},
-			}},
-
-			notValid: true,
-			reason:   InvalidSpecReason,
-			message:  InvalidDurationMessage,
-		},
-	}
-	for i := range tests {
-		test := tests[i]
-		t.Run(test.name, func(t *testing.T) {
-			ro := &v1alpha1.Rollout{
-				Spec: v1alpha1.RolloutSpec{
-					Selector: &metav1.LabelSelector{
-						MatchLabels: map[string]string{"key": "value"},
-					},
-					Strategy: v1alpha1.RolloutStrategy{
-						Canary: &v1alpha1.CanaryStrategy{
-							MaxUnavailable: test.maxUnavailable,
-							MaxSurge:       test.maxSurge,
-							Steps:          test.steps,
-						},
-					},
-				},
-			}
-			cond := VerifyRolloutSpec(ro, nil)
-			if test.notValid {
-				assert.Equal(t, v1alpha1.InvalidSpec, cond.Type)
-				assert.Equal(t, test.reason, cond.Reason)
-				assert.Equal(t, test.message, cond.Message)
-			} else {
-				assert.Nil(t, cond)
-			}
-		})
-	}
-}
-
-func TestInvalidMaxSurgeMaxUnavailable(t *testing.T) {
-	r := func(maxSurge, maxUnavailable intstr.IntOrString) *v1alpha1.Rollout {
-		return &v1alpha1.Rollout{
-			Spec: v1alpha1.RolloutSpec{
-				Strategy: v1alpha1.RolloutStrategy{
-					Canary: &v1alpha1.CanaryStrategy{
-						MaxSurge:       &maxSurge,
-						MaxUnavailable: &maxUnavailable,
-					},
-				},
-			},
-		}
-	}
-	assert.True(t, invalidMaxSurgeMaxUnavailable(r(intstr.FromInt(0), intstr.FromInt(0))))
-	assert.True(t, invalidMaxSurgeMaxUnavailable(r(intstr.FromString("0"), intstr.FromInt(0))))
-	assert.True(t, invalidMaxSurgeMaxUnavailable(r(intstr.FromString("0%"), intstr.FromInt(0))))
-	assert.True(t, invalidMaxSurgeMaxUnavailable(r(intstr.FromInt(0), intstr.FromString("0"))))
-	assert.True(t, invalidMaxSurgeMaxUnavailable(r(intstr.FromInt(0), intstr.FromString("0%"))))
-	assert.True(t, invalidMaxSurgeMaxUnavailable(r(intstr.FromString("0"), intstr.FromString("0"))))
-	assert.True(t, invalidMaxSurgeMaxUnavailable(r(intstr.FromString("0%"), intstr.FromString("0%"))))
-
-}
-
-func TestHasRevisionHistoryLimit(t *testing.T) {
-	r := &v1alpha1.Rollout{}
-	assert.False(t, HasRevisionHistoryLimit(r))
-	int32Value := int32(math.MaxInt32)
-	r.Spec.RevisionHistoryLimit = &int32Value
-	assert.False(t, HasRevisionHistoryLimit(r))
-	int32Value = int32(1)
-	r.Spec.RevisionHistoryLimit = &int32Value
-	assert.True(t, HasRevisionHistoryLimit(r))
 }
 
 func TestRolloutProgressing(t *testing.T) {
@@ -842,53 +614,4 @@ func TestComputeStepHash(t *testing.T) {
 	assert.NotEqual(t, baseline, roWithDiffStepsHash)
 	assert.Equal(t, baseline, roWithSameStepsHash)
 	assert.NotEqual(t, baseline, roNoStepsHash)
-}
-
-func TestInvalidAntiAffinity(t *testing.T) {
-	affinity := v1alpha1.AntiAffinity{}
-	reason, message := invalidAntiAffinity(affinity, "BlueGreen")
-	expectedMsg := fmt.Sprintf(MissingFieldMessage, ".Spec.Strategy.BlueGreen.AntiAffinity.PreferredDuringSchedulingIgnoredDuringExecution or .Spec.Strategy.BlueGreen.AntiAffinity.RequiredDuringSchedulingIgnoredDuringExecution")
-	assert.Equal(t, InvalidSpecReason, reason)
-	assert.Equal(t, expectedMsg, message)
-
-	affinity = v1alpha1.AntiAffinity{
-		RequiredDuringSchedulingIgnoredDuringExecution:  &v1alpha1.RequiredDuringSchedulingIgnoredDuringExecution{},
-		PreferredDuringSchedulingIgnoredDuringExecution: &v1alpha1.PreferredDuringSchedulingIgnoredDuringExecution{Weight: 1},
-	}
-	reason, message = invalidAntiAffinity(affinity, "Canary")
-	assert.Equal(t, InvalidSpecReason, reason)
-	assert.Equal(t, "Multiple Anti-Affinity Strategies can not be listed", message)
-}
-
-func TestVerifyRolloutSpecAntiAffinity(t *testing.T) {
-	affinity := &v1alpha1.AntiAffinity{}
-	invalidRollout := &v1alpha1.Rollout{
-		Spec: v1alpha1.RolloutSpec{
-			Selector: &metav1.LabelSelector{
-				MatchLabels: map[string]string{"key": "value"},
-			},
-			Strategy: v1alpha1.RolloutStrategy{
-				BlueGreen: &v1alpha1.BlueGreenStrategy{
-					PreviewService: "preview",
-					ActiveService:  "active",
-					AntiAffinity:   affinity,
-				},
-			},
-		},
-	}
-	cond := VerifyRolloutSpec(invalidRollout, nil)
-	message := fmt.Sprintf(MissingFieldMessage, ".Spec.Strategy.BlueGreen.AntiAffinity.PreferredDuringSchedulingIgnoredDuringExecution or .Spec.Strategy.BlueGreen.AntiAffinity.RequiredDuringSchedulingIgnoredDuringExecution")
-	assert.Equal(t, InvalidSpecReason, cond.Reason)
-	assert.Equal(t, message, cond.Message)
-
-	invalidRollout.Spec.Strategy.BlueGreen = nil
-	invalidRollout.Spec.Strategy.Canary = &v1alpha1.CanaryStrategy{AntiAffinity: &v1alpha1.AntiAffinity{
-		PreferredDuringSchedulingIgnoredDuringExecution: &v1alpha1.PreferredDuringSchedulingIgnoredDuringExecution{
-			Weight: 1,
-		},
-		RequiredDuringSchedulingIgnoredDuringExecution: &v1alpha1.RequiredDuringSchedulingIgnoredDuringExecution{},
-	}}
-	cond = VerifyRolloutSpec(invalidRollout, nil)
-	assert.Equal(t, InvalidSpecReason, cond.Reason)
-	assert.Equal(t, "Multiple Anti-Affinity Strategies can not be listed", cond.Message)
 }
